renamed kuzzleUser.js -> User.js
